+++ randovania/layout/description_migration.py
@@ -6,7 +6,7 @@ from randovania.game_description import migration_data
 from randovania.games.game import RandovaniaGame
 from randovania.lib import migration_lib
 
-CURRENT_VERSION = 9
+CURRENT_VERSION = 10
 
 
 def _migrate_v1(json_dict: dict) -> dict:
@@ -186,6 +186,12 @@ def _migrate_v8(json_dict: dict) -> dict:
     return json_dict
 
 
+def _migrate_v9(json_dict: dict) -> dict:
+    json_dict["info"]["has_spoiler"] = "game_modifications" in json_dict
+
+    return json_dict
+
+
 _MIGRATIONS = {
     1: _migrate_v1,  # v2.2.0-6-gbfd37022
     2: _migrate_v2,  # v2.4.2-16-g735569fd
@@ -195,6 +201,7 @@ _MIGRATIONS = {
     6: _migrate_v6,
     7: _migrate_v7,  # v3.3.0dev721
     8: _migrate_v8,
+    9: _migrate_v9,
 }