+++ randovania/server/game_session.py
@@ -351,6 +351,58 @@ def _change_password(sio: ServerApp, session: GameSession, password: str):
     _add_audit_entry(sio, session, f"Changed password")
 
 
+def _change_title(sio: ServerApp, session: GameSession, title: str):
+    _verify_has_admin(sio, session.id, None)
+
+    old_name = session.name
+    session.name = title
+    logger().info(f"{_describe_session(session)}: Changed name from {old_name}.")
+    session.save()
+    _add_audit_entry(sio, session, f"Changed name from {old_name} to {title}")
+
+
+def _duplicate_session(sio: ServerApp, session: GameSession, new_title: str):
+    _verify_has_admin(sio, session.id, None)
+
+    current_user = sio.get_current_user()
+    _add_audit_entry(sio, session, f"Duplicated session as {new_title}")
+
+    with database.db.atomic():
+        new_session: GameSession = GameSession.create(
+            name=new_title,
+            password=session.password,
+            creator=current_user,
+            layout_description_json=session.layout_description_json,
+            seed_hash=session.seed_hash,
+            dev_features=session.dev_features,
+        )
+        for preset in session.presets:
+            assert isinstance(preset, GameSessionPreset)
+            GameSessionPreset.create(
+                session=new_session,
+                row=preset.row,
+                preset=preset.preset,
+            )
+        GameSessionMembership.create(
+            user=current_user,
+            session=new_session,
+            row=None, admin=True,
+            connection_state="Offline",
+        )
+        GameSessionAudit.create(
+            session=new_session,
+            user=current_user,
+            message=f"Duplicated from {session.name}",
+        )
+
+
+def _get_permalink(sio: ServerApp, session: GameSession) -> str:
+    _verify_has_admin(sio, session.id, None)
+    _add_audit_entry(sio, session, "Requested permalink")
+
+    return session.layout_description.permalink.as_base64_str
+
+
 def game_session_admin_session(sio: ServerApp, session_id: int, action: str, arg):
     action: SessionAdminGlobalAction = SessionAdminGlobalAction(action)
     session: database.GameSession = database.GameSession.get_by_id(session_id)
@@ -385,9 +437,18 @@ def game_session_admin_session(sio: ServerApp, session_id: int, action: str, arg
     elif action == SessionAdminGlobalAction.CHANGE_PASSWORD:
         _change_password(sio, session, arg)
 
+    elif action == SessionAdminGlobalAction.CHANGE_TITLE:
+        _change_title(sio, session, arg)
+
+    elif action == SessionAdminGlobalAction.DUPLICATE_SESSION:
+        return _duplicate_session(sio, session, arg)
+
     elif action == SessionAdminGlobalAction.DELETE_SESSION:
         logger().info(f"{_describe_session(session)}: Deleting session.")
         session.delete_instance(recursive=True)
+
+    elif action == SessionAdminGlobalAction.REQUEST_PERMALINK:
+        return _get_permalink(sio, session)
 
     _emit_session_meta_update(session)