+++ test/gui/preset_settings/test_preset_elevator.py
@@ -27,4 +27,4 @@ def test_on_preset_changed(skip_qtbot, preset_manager, game):
     # Assert
     num_areas = len(TeleporterTargetList.areas_list(preset.game))
     assert len(window._elevator_target_for_area) == num_areas
-    assert window.tab_title == "Teleporters" if preset.game == RandovaniaGame.METROID_PRIME_CORRUPTION else "Elevators"
+    assert "Elevators"